Marked unsupported field for supported APIs
For all the supported API, fields which are not supported by Nutanix Objects are
marked with isNtnxSupported tag set to false. It is done in top down manner, i.e,
start from the operation level and wheneven a field is seen as unsupported, the
field, and all its sub-fields should be skipped while generating the documentation.

Python script graph.py is used to list out all the child fields given an operation